Fix pip unless statement
Using the `pip search` function any python package which includes the
package name will be returned and if it's installed the `python::pip`
class will skip the `exec`.  One example of this is the docker-compose
package and the docker-py package.  Since docker-py includes
"docker-compose" in the description a pip install of `docker-compose`
after docker-py will not run.

This should resolve this by fist looking if the package is installed.
Then looking if it is up to date.